removed func from script
